TestProject/requirements.txt
amqp==2.3.2
billiard==3.5.0.4
celery==4.2.1
certifi==2018.8.24
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==1.11
django-cors-headers==2.4.0
django-crispy-forms==1.7.2
django-filter==2.0.0
django-formtools==2.1
django-jet==1.0.7
django-redis==4.9.0
django-reversion==3.0.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
drf-extensions==0.4.0
future==0.16.0
httplib2==0.9.2
idna==2.7
itypes==1.1.0
Jinja2==2.10
kombu==4.2.1
Markdown==3.0.1
MarkupSafe==1.0
mysqlclient==1.3.13
Pillow==5.3.0
pycryptodome==3.6.6
PyJWT==1.6.4
pytz==2018.5
raven==6.9.0
redis==2.10.6
requests==2.19.1
six==1.11.0
uritemplate==3.0.0
urllib3==1.23
uwsgi
vine==1.1.4
XlsxWriter==1.1.1
xlwt==1.3.0
